separate panel from overlay, merge
